Changes to be committed: 	modified:   benchmark/quantum_integration.py 	renamed:    benchmark/jsons_2/cqpeae_configuration.json -> benchmark/test/json_tests/cqpeae_configuration.json 	renamed:    benchmark/jsons_2/iqae_configuration.json -> benchmark/test/json_tests/iqae_configuration.json 	renamed:    benchmark/jsons_2/iqpeae_configuration.json -> benchmark/test/json_tests/iqpeae_configuration.json 	renamed:    benchmark/jsons_2/mlae_configuration.json -> benchmark/test/json_tests/mlae_configuration.json 	renamed:    benchmark/jsons_2/rqae_configuration.json -> benchmark/test/json_tests/rqae_configuration.json 	new file:   benchmark/test/quantum_integration_test.py
